add condition, if image exists, do not download the image any more

+++ douban/douban/pipelines.py
@@ -135,13 +135,15 @@ class MoviePipeline(object):
         name = url.split('/')[-1]
         fname = mid + "_" + name
         fpath =os.path.join(self.image_dir, fname) 
-        urllib.urlretrieve(url, fpath)
+        if os.exists(fpath) == False:
+        	urllib.urlretrieve(url, fpath)
 
     def save_photo(self, mid, url):
         name = url.split('/')[-1]
         fname = mid + "_" + name
-        fpath =os.path.join(self.photo_dir, fname) 
-        urllib.urlretrieve(url, fpath)
+        fpath =os.path.join(self.photo_dir, fname)
+        if os.exists(fpath) == False:
+        	urllib.urlretrieve(url, fpath)
 
     def process_actor(self, item):
         self.dbpool.runInteraction(self.actor_insert, item)
